+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphand/client
 
+## 1.4.1
+
+### Patch Changes
+
+- add buildMediaUrl on graphand client
+
 ## 1.4.0
 
 ### Minor Changes